Tests/LibWeb/Text/expected/css/range-descriptor-calculated-value.txt

@counter-style valid { range: sign(1em - 1px) calc(1 + sign(1em - 1px)); }
@counter-style invalid { }
